+++ src/generator.rs
@@ -90,6 +90,17 @@ impl Generator
    */
   pub fn generate(&self) -> bool
   {
+    match crate::checks::sqlite_check()
+    {
+      crate::checks::Status::Error(msg) =>
+      {
+        eprintln!("{}", msg);
+        return false;
+      },
+      crate::checks::Status::Warn(msg) => println!("Warning: {}", msg),
+      _ => ()
+    }
+
     let db = Database::new(&self.config.db_path);
     let db = match db
     {